# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MultiLiveTV API
  version: 1.0.0
  description: VOD aggregation API for Apple TV / iPad / Android clients
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /vod/list:
    get:
      summary: VOD list by category
      tags: [VOD]
      parameters:
        - name: sourceId
          in: query
          schema: { type: integer }
        - name: pg
          in: query
          schema: { type: integer, default: 1 }
        - name: t
          in: query
          description: Category type id (single-source)
          schema: { type: integer }
        - name: cat
          in: query
          description: Unified category slug (multi-source; takes precedence over t)
          schema: { type: string }
      responses:
        "200":
          description: OK
  /vod/detail:
    get:
      summary: Merged VOD detail with play sources
      tags: [VOD]
      parameters:
        - name: sourceId
          in: query
          required: true
          schema: { type: integer }
        - name: ids
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT